# server/openapi.yaml
# 프론트엔드 개발자에게 보여주는 API 명세.
# docs/02-api.md 와 1:1 로 맞춘다 — 둘이 어긋나면 02-api.md 가 기준이고 이 파일을 고친다.
#
# 3단계 기준이라 인증이 x-user-id 헤더다. 6단계에서 Authorization: Bearer <JWT> 로 바뀐다.
openapi: 3.1.0

info:
  title: 게시판 API
  version: 0.6.0
  description: |
    게시판 REST API. 리소스는 users · posts · comments 셋이다.

    ### 응답 형태는 세 가지뿐이다
    - **단건**: 객체 그대로. 봉투(`{ data: ... }`)를 씌우지 않는다.
    - **목록**: `{ items, pageInfo }`.
    - **에러**: `{ error: { code, message, details } }`. 모든 4xx·5xx 가 이 형태 하나다.

    ### 프론트엔드가 알아야 할 것
    - 에러 분기는 `message` 가 아니라 **`error.code`** 로 한다. `message` 는 사람이 읽는 문구라 바뀐다.
    - `401` 은 로그인 화면으로 보낸다. `403` 은 보내지 않는다 — 다시 로그인해도 해결되지 않는다.
    - 목록은 커서 페이지네이션이다. `pageInfo.nextCursor` 를 다음 요청의 `cursor` 에 그대로 넣는다.
      `hasNext` 가 `false` 면 끝이다. 페이지 번호로 점프하는 방법은 없다.
    - 글 상세(`GET /posts/{id}`)는 작성자와 **댓글 첫 20개**를 함께 준다. 상세 화면을 그리는 데 호출 한 번이면 된다.
    - 현재 단계에서는 `sort` · `q` · `authorId` 쿼리 파라미터가 아직 동작하지 않는다(보내면 무시된다).

servers:
  - url: http://localhost:4000
    description: 로컬 개발

tags:
  - name: auth
    description: 가입 · 로그인 · 내 정보
  - name: posts
    description: 글
  - name: comments
    description: 댓글
  - name: system
    description: 헬스체크

security:
  - bearerAuth: []

paths:
  /health:
    get:
      tags: [system]
      summary: 살아 있는지 확인
      security: []
      responses:
        "200":
          description: 정상
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, const: true }

  /auth/signup:
    post:
      tags: [auth]
      summary: 가입
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignupRequest" }
            example:
              email: kim@example.com
              password: password123
              nickname: kim
      responses:
        "201":
          description: 가입 성공. 비밀번호는 응답에 실리지 않는다.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: 이미 가입된 이메일
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error: { code: EMAIL_ALREADY_EXISTS, message: 이미 가입된 이메일입니다., details: null }
        "422": { $ref: "#/components/responses/Unprocessable" }

  /auth/login:
    post:
      tags: [auth]
      summary: 로그인, 토큰 발급
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
            example: { email: kim@example.com, password: password123 }
      responses:
        "200":
          description: |
            로그인 성공. `token` 을 이후 요청의 `Authorization: Bearer <token>` 헤더에 넣는다.
            함께 `httpOnly` 쿠키로 리프레시 토큰이 내려간다 — 자바스크립트로는 읽을 수 없다.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: |
            이메일 또는 비밀번호가 맞지 않음.
            **없는 이메일과 틀린 비밀번호를 구분하지 않는다** — 구분하면 가입 여부가 새어 나간다.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error: { code: INVALID_CREDENTIALS, message: 이메일 또는 비밀번호가 맞지 않습니다., details: null }

  /auth/refresh:
    post:
      tags: [auth]
      summary: 액세스 토큰 재발급
      description: |
        `httpOnly` 쿠키(`refresh_token`)를 읽어 새 액세스 토큰을 준다.
        **요청 바디도 Authorization 헤더도 필요 없다** — 쿠키가 브라우저에서 자동으로 실려 간다.
        성공하면 리프레시 토큰도 새것으로 갈아 끼운다.
      security: []
      responses:
        "200":
          description: 재발급 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401":
          description: 쿠키가 없거나, 만료됐거나, 리프레시 토큰이 아님
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error: { code: UNAUTHENTICATED, message: 리프레시 토큰이 없습니다., details: null }

  /me:
    get:
      tags: [auth]
      summary: 내 정보
      responses:
        "200":
          description: 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthenticated" }

  /posts:
    get:
      tags: [posts]
      summary: 글 목록
      security: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: sort
          in: query
          description: "`createdAt:desc` 등. **7단계에서 구현된다 — 지금은 무시된다.**"
          schema: { type: string, enum: ["createdAt:desc", "createdAt:asc"], default: "createdAt:desc" }
        - name: q
          in: query
          description: "제목 검색. **7단계에서 구현된다 — 지금은 무시된다.**"
          schema: { type: string }
        - name: authorId
          in: query
          description: "작성자 필터. **7단계에서 구현된다 — 지금은 무시된다.**"
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: 최신순 목록. 항목에 `body` 는 없다 — 상세에서만 준다.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PostPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
    post:
      tags: [posts]
      summary: 글 작성
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePostRequest" }
            example: { title: 첫 글, body: 본문이다. }
      responses:
        "201":
          description: 작성 성공. `Location` 헤더에 새 글의 경로가 담긴다.
          headers:
            Location:
              description: 새 글의 경로
              schema: { type: string, examples: ["/posts/9f1c…"] }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PostDetail" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "422": { $ref: "#/components/responses/Unprocessable" }

  /posts/{id}:
    parameters:
      - $ref: "#/components/parameters/PostId"
    get:
      tags: [posts]
      summary: 글 상세 (작성자 + 댓글 첫 20개 포함)
      security: []
      responses:
        "200":
          description: 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PostDetail" }
        "404": { $ref: "#/components/responses/PostNotFound" }
    patch:
      tags: [posts]
      summary: 글 수정 (소유자만)
      description: 부분 수정이다. 보낸 필드만 바뀌고 안 보낸 필드는 그대로 남는다.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePostRequest" }
            example: { title: 제목만 고친다 }
      responses:
        "200":
          description: 수정된 글
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PostDetail" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/PostNotFound" }
        "422": { $ref: "#/components/responses/Unprocessable" }
    delete:
      tags: [posts]
      summary: 글 삭제 (소유자만)
      description: |
        딸린 댓글도 함께 지워진다.
        **같은 요청을 두 번 보내면 두 번째는 `404`** 다 — 없는 글에 대한 응답을 메서드마다
        다르게 만들지 않기 위한 결정이다. 낙관적 삭제에서 `404` 는 "이미 없음" 이므로 성공으로 쳐도 된다.
      responses:
        "204": { description: 삭제 성공. 바디가 없다. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/PostNotFound" }

  /posts/{id}/comments:
    parameters:
      - $ref: "#/components/parameters/PostId"
    get:
      tags: [comments]
      summary: 댓글 목록
      description: 글 목록과 달리 **오래된 것부터** 온다. 읽는 순서가 대화 순서이기 때문이다.
      security: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommentPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/PostNotFound" }
    post:
      tags: [comments]
      summary: 댓글 작성
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCommentRequest" }
            example: { body: 첫 댓글이다. }
      responses:
        "201":
          description: |
            작성 성공. `Location` 헤더는 없다 — 댓글 단건을 조회하는 경로가 이 API 에 없다.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comment" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/PostNotFound" }
        "422": { $ref: "#/components/responses/Unprocessable" }

  /comments/{id}:
    delete:
      tags: [comments]
      summary: 댓글 삭제 (소유자만)
      description: 댓글 작성자만 지운다. 글 작성자가 남의 댓글을 지우는 권한은 없다.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: 삭제 성공. 바디가 없다. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: 그런 댓글이 없음
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error: { code: COMMENT_NOT_FOUND, message: 댓글을 찾을 수 없습니다., details: null }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `POST /auth/login` 이 주는 **액세스 토큰(JWT, 15분)** 을 넣는다.
        페이로드에는 `sub`(유저 id)와 `exp` 만 있다 — 토큰은 서명된 것이지 암호화된 것이 아니라
        누구나 읽을 수 있기 때문이다.

        만료되면 `401` 이 온다. 그때 `POST /auth/refresh` 를 부르면
        `httpOnly` 쿠키에 담긴 리프레시 토큰으로 새 액세스 토큰을 받는다.

  parameters:
    Limit:
      name: limit
      in: query
      description: 한 번에 받을 개수
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
    Cursor:
      name: cursor
      in: query
      description: 앞 응답의 `pageInfo.nextCursor` 를 그대로 넣는다. 없으면 첫 페이지다.
      schema: { type: string }
    PostId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }

  responses:
    BadRequest:
      description: 형식이 틀림 (타입 불일치, 없는 값, 범위 밖)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: VALIDATION_FAILED, message: "title, body 는 문자열이어야 합니다.", details: null }
    Unprocessable:
      description: 형식은 맞지만 값이 규칙 위반 (길이 초과, 빈 문자열)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: VALIDATION_FAILED
              message: title 은 1자 이상 200자 이하여야 합니다.
              details: [{ path: title, rule: too_big }]
    Unauthenticated:
      description: 누군지 모름. 토큰이 없거나 만료됐거나 서명이 안 맞음. **로그인 화면으로 보낸다.**
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: UNAUTHENTICATED, message: 로그인이 필요합니다., details: null }
    Forbidden:
      description: |
        누군지는 아는데 권한이 없음. 남의 글·댓글을 고치거나 지우려 할 때.
        **로그인 화면으로 보내지 않는다** — 다시 로그인해도 해결되지 않는다.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: FORBIDDEN, message: 내 글만 수정할 수 있습니다., details: null }
    PostNotFound:
      description: 그런 글이 없음
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: POST_NOT_FOUND, message: 글을 찾을 수 없습니다., details: null }

  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              description: 프론트엔드가 분기하는 값. 아래 목록 밖의 값은 오지 않는다.
              enum:
                - VALIDATION_FAILED
                - EMAIL_ALREADY_EXISTS
                - INVALID_CREDENTIALS
                - UNAUTHENTICATED
                - FORBIDDEN
                - POST_NOT_FOUND
                - COMMENT_NOT_FOUND
                - ROUTE_NOT_FOUND
                - INTERNAL_ERROR
            message:
              type: string
              description: 사람이 읽는 문구. 이 값으로 분기하지 않는다.
            details:
              description: 검증 실패일 때만 채워진다.
              type: [array, "null"]
              items:
                type: object
                properties:
                  path: { type: string }
                  rule: { type: string }

    PageInfo:
      type: object
      required: [nextCursor, hasNext]
      properties:
        nextCursor:
          type: [string, "null"]
          description: 다음 요청의 `cursor` 에 넣는 값. 마지막 페이지에서는 `null` 이다.
        hasNext: { type: boolean }

    Author:
      type: object
      required: [id, nickname]
      properties:
        id: { type: string, format: uuid }
        nickname: { type: string }

    User:
      type: object
      required: [id, email, nickname, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        nickname: { type: string }
        createdAt: { type: string, format: date-time }

    PostListItem:
      type: object
      description: 목록 항목. `body` 가 없다.
      required: [id, title, author, commentCount, createdAt]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        author: { $ref: "#/components/schemas/Author" }
        commentCount: { type: integer }
        createdAt: { type: string, format: date-time }

    PostDetail:
      type: object
      required: [id, title, body, author, commentCount, comments, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        body: { type: string }
        author: { $ref: "#/components/schemas/Author" }
        commentCount: { type: integer }
        comments:
          $ref: "#/components/schemas/CommentPage"
          description: 댓글 첫 20개. 더 있으면 `pageInfo.hasNext` 가 `true` 다.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    Comment:
      type: object
      required: [id, body, author, createdAt]
      properties:
        id: { type: string, format: uuid }
        body: { type: string }
        author: { $ref: "#/components/schemas/Author" }
        createdAt: { type: string, format: date-time }

    PostPage:
      type: object
      required: [items, pageInfo]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/PostListItem" }
        pageInfo: { $ref: "#/components/schemas/PageInfo" }

    CommentPage:
      type: object
      required: [items, pageInfo]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Comment" }
        pageInfo: { $ref: "#/components/schemas/PageInfo" }

    SignupRequest:
      type: object
      required: [email, password, nickname]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, description: 8자 이상 }
        nickname: { type: string }

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }

    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: 액세스 토큰(JWT, 15분).
        user: { $ref: "#/components/schemas/User" }

    CreatePostRequest:
      type: object
      required: [title, body]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        body: { type: string, minLength: 1 }

    UpdatePostRequest:
      type: object
      description: 최소 한 필드는 보낸다. 보낸 필드만 바뀐다.
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        body: { type: string, minLength: 1 }

    CreateCommentRequest:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1, maxLength: 1000 }